move daemon mode judge helper to a new source file
